+++ config/test.exs
@@ -7,4 +7,7 @@ config :forge, Forge.Repo,
   hostname: "localhost",
   pool: Ecto.Adapters.SQL.Sandbox
 
+config :forge, :start_repo, true
+config :forge, :api_server, enabled: false
+
 config :logger, level: :warning