+++ lib/forge/application.ex
@@ -10,17 +10,24 @@ defmodule Forge.Application do
     # Attach telemetry handlers before starting supervision tree
     attach_telemetry_handlers()
 
-    children = [
-      Forge.Repo,
-      {Task.Supervisor, name: Forge.MeasurementTaskSupervisor}
-      # Starts a worker by calling: Forge.Worker.start_link(arg)
-      # {Forge.Worker, arg}
-    ]
+    children =
+      []
+      |> maybe_child(Application.get_env(:forge, :start_repo, true), Forge.Repo)
+      |> maybe_child(true, {Task.Supervisor, name: Forge.MeasurementTaskSupervisor})
+      |> maybe_child(api_enabled?(), Forge.API.Server)
 
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: Forge.Supervisor]
     Supervisor.start_link(children, opts)
+  end
+
+  defp maybe_child(children, true, child), do: children ++ [child]
+  defp maybe_child(children, _flag, _child), do: children
+
+  defp api_enabled? do
+    config = Application.get_env(:forge, :api_server, [])
+    Keyword.get(config, :enabled, false)
   end
 
   defp attach_telemetry_handlers do